@@ -169,10 +169,10 @@
     "latest-version": "^5.1.0",
     "merge": "^1.2.0",
     "npm-install-version": "^6.0.2",
-    "react": "16.13.1",
-    "react-beautiful-dnd": "^13.0.0",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "16.13.1",
+    "react": "^17.0.2",
+    "react-beautiful-dnd": "^13.1.0",
+    "react-bootstrap": "^1.6.4",
+    "react-dom": "^17.0.2",
     "selenium": "^2.20.0",
     "signale": "^1.4.0",
     "string-similarity": "^4.0.4",
@@ -290,10 +291,10 @@
     "tap-spec": "^5.0.0",
     "tape": "^4.13.3",
     "terser-webpack-plugin": "^4.2.3",
-    "ts-jest": "25.2.1",
+    "ts-jest": "^27.0.5",
     "ts-node": "^7.0.1",
     "tslint": "~6.0.0",
-    "typescript": "~3.8.3",
+    "typescript": "^4.4.3",
     "uglify-js": "^2.8.16",
     "vm-browserify": "0.0.4",
     "watchify": "^3.9.0",
