@@ -108,7 +108,7 @@ function sourceLocationFromAstNode (astNode) {
  * @param {Object} ast - ast given by the compilation result
  */
 export function findNodeAtInstructionIndex (astNodeType, instIndex, sourceMap, ast) {
-  const sourceLocation = this.atIndex(instIndex, sourceMap)
+  const sourceLocation = atIndex(instIndex, sourceMap)
   return findNodeAtSourceLocation(astNodeType, sourceLocation, ast)
 }
 
