@@ -3,8 +3,8 @@ import { ethers } from 'ethers'
 
 export function makeFullTypeDefinition (typeDef) {
   if (typeDef && typeDef.type.indexOf('tuple') === 0 && typeDef.components) {
-    const innerTypes = typeDef.components.map((innerType) => { return this.makeFullTypeDefinition(innerType) })
-    return `tuple(${innerTypes.join(',')})${this.extractSize(typeDef.type)}`
+    const innerTypes = typeDef.components.map((innerType) => { return makeFullTypeDefinition(innerType) })
+    return `tuple(${innerTypes.join(',')})${extractSize(typeDef.type)}`
   }
   return typeDef.type
 }
@@ -19,7 +19,7 @@ export function encodeParams (funABI, args) {
       if (type === 'bool' && args[i] === 'false') {
         args[i] = false
       }
-      types.push(type.indexOf('tuple') === 0 ? this.makeFullTypeDefinition(funABI.inputs[i]) : type)
+      types.push(type.indexOf('tuple') === 0 ? makeFullTypeDefinition(funABI.inputs[i]) : type)
       if (args.length < types.length) {
         args.push('')
       }
@@ -107,7 +107,7 @@ export function getFunction (abi, fnName) {
     const fn = abi[i]
     if (fn.type === 'function' && fnName === fn.name + '(' + fn.inputs.map((value) => {
       if (value.components) {
-        const fullType = this.makeFullTypeDefinition(value)
+        const fullType = makeFullTypeDefinition(value)
         return fullType.replace(/tuple/g, '') // return of makeFullTypeDefinition might contain `tuple`, need to remove it cause `methodIdentifier` (fnName) does not include `tuple` keyword
       } else {
         return value.type
