@@ -129,7 +129,11 @@ export class Web3VmProvider {
     this.storageCache[this.processingHash] = {}
     if (data.to) {
       try {
-        const storage = await this.vm.stateManager.dumpStorage(data.to)
+        // dumpStorage throws error as 'Missing Node in DB'
+        // This can be uncommented once that error is handled
+        // https://github.com/ethereum/remix-project/issues/1644
+        // const storage = await this.vm.stateManager.dumpStorage(data.to)
+        const storage = {}
         this.storageCache[this.processingHash][tx['to']] = storage
         this.lastProcessedStorageTxHash[tx['to']] = this.processingHash
       } catch (e) {
@@ -245,7 +249,11 @@ export class Web3VmProvider {
         if (!this.storageCache[this.processingHash][this.processingAddress]) {
           const account = Address.fromString(this.processingAddress)
           try {
-            const storage = await this.vm.stateManager.dumpStorage(account)
+            // dumpStorage throws error as 'Missing Node in DB'
+            // This can be uncommented once that error is handled
+            // https://github.com/ethereum/remix-project/issues/1644
+            // const storage = await this.vm.stateManager.dumpStorage(account)
+            const storage = {}
             this.storageCache[this.processingHash][this.processingAddress] = storage
             this.lastProcessedStorageTxHash[this.processingAddress] = this.processingHash
           } catch (e) {
