@@ -11,12 +11,12 @@ interface SetPluginOptionType {
 }
 
 export class RemixEngine extends Engine {
-  event: EventEmitter;
+  event: EventEmitter
   setPluginOption ({ name, kind }) : SetPluginOptionType
   onRegistration (plugin) : void
 }
 
-export function isNative(name: any): any;
+export function isNative(name: any): any
 /**
  * Checks if plugin caller 'from' is allowed to activate plugin 'to'
  * The caller can have 'canActivate' as a optional property in the plugin profile.
@@ -27,33 +27,33 @@ export function isNative(name: any): any;
  * @param {any, any}
  * @returns {boolean}
  */
-export function canActivate(from: any, to: any): boolean;
+export function canActivate(from: any, to: any): boolean
 export class RemixAppManager extends PluginManager {
-  constructor();
-    event: EventEmitter;
-    pluginsDirectory: string;
-    pluginLoader: PluginLoader;
-    permissionHandler: PermissionHandler;
-    getAll(): import('@remixproject/plugin-utils').Profile<any>[];
-    getIds(): string[];
-    isDependent(name: any): any;
-    isRequired(name: any): any;
-    registeredPlugins(): Promise<any>;
-    turnPluginOn(name: string | string[]);
-    turnPluginOff(name: string);
+  constructor()
+  event: EventEmitter
+  pluginsDirectory: string
+  pluginLoader: PluginLoader
+  permissionHandler: PermissionHandler
+  getAll(): import('@remixproject/plugin-utils').Profile<any>[]
+  getIds(): string[]
+  isDependent(name: any): any
+  isRequired(name: any): any
+  registeredPlugins(): Promise<any>
+  turnPluginOn(name: string | string[])
+  turnPluginOff(name: string)
 }
 
 export class PluginManagerSettings {
-  openDialog(): void;
-  permissions: any;
-  currentSetting: any;
-  onValidation(): void;
+  openDialog(): void
+  permissions: any
+  currentSetting: any
+  onValidation(): void
   /** Clear one permission from a plugin */
-  clearPersmission(from: string, to: string, method: string): void;
+  clearPersmission(from: string, to: string, method: string): void
   /** Clear all persmissions from a plugin */
-  clearAllPersmission(to: string): void;
-  settings(): any;
-  render(): any;
+  clearAllPersmission(to: string): void
+  settings(): any
+  render(): any
 }
 
 export type PluginPermissions = {
@@ -94,7 +94,7 @@ export class PluginManagerComponent extends ViewPlugin extends Plugin implements
 }
 
 // eslint-disable-next-line no-use-before-define
-export = LocalPlugin;
+export = LocalPlugin
 declare class LocalPlugin {
   /**
      * Open a modal to create a local plugin
@@ -102,34 +102,34 @@ declare class LocalPlugin {
      * @returns {Promise<{api: any, profile: any}>} A promise with the new plugin profile
      */
   open(plugins: any[]): Promise<{
-        api: any;
-        profile: any;
-    }>;
+        api: any
+        profile: any
+    }>
 
-    profile: any;
+    profile: any
     /**
      * Create the object to add to the plugin-list
      */
-    create(): any;
+    create(): any
     updateName({ target }: {
-        target: any;
-    }): void;
+        target: any
+    }): void
 
     updateUrl({ target }: {
-        target: any;
-    }): void;
+        target: any
+    }): void
 
     updateDisplayName({ target }: {
-        target: any;
-    }): void;
+        target: any
+    }): void
 
-    updateProfile(key: any, e: any): void;
+    updateProfile(key: any, e: any): void
     updateMethods({ target }: {
-        target: any;
-    }): void;
+        target: any
+    }): void
 
     /** The form to create a local plugin */
-    form(): any;
+    form(): any
 }
 
 export interface PluginManagerContextProviderProps {
@@ -146,12 +146,12 @@ export interface RemixUiPluginManagerProps {
  *  (localStorage, queryParams)
  **/
 declare class PluginLoader {
-  get currentLoader(): any;
-    donotAutoReload: string[];
-    loaders: {};
-    current: string;
-    set(plugin: any, actives: any): void;
-    get(): any;
+  get currentLoader(): any
+    donotAutoReload: string[]
+    loaders: {}
+    current: string
+    set(plugin: any, actives: any): void
+    get(): any
 }
 
 export type PluginManagerSettings = {
